qemu functions in own files, hidable with pre-proc
- Hide qemu I/O in their own header file with macros if NODEBUG.
- Move qemu functions to their own file: libcommon/qemu_debug.c
- Harmonise implementation with firmware.
