+++ README.md
@@ -116,8 +116,8 @@ debug port on 0xfe00\_1000 (`TK1_MMIO_QEMU_DEBUG`). Anything written
 there will be printed as a character by qemu on the console.
 
 `qemu_putchar()`, `qemu_puts()`, `qemu_putinthex()`, `qemu_hexdump()`
-and friends (see `libcommon/lib.[ch]`) use this debug port to print
-stuff.
+and friends (see `libcommon/lib.h` and `libcommon/qemu_debug.c`) use
+this debug port to print stuff.
 
 `libcommon` is compiled with no debug output by default. Rebuild
 `libcommon` without `-DNODEBUG` to get the debug output.
