+++ libcommon/lib.c
@@ -5,131 +5,6 @@
 #include <tk1_mem.h>
 #include <types.h>
 
-#ifdef NODEBUG
-int qemu_putchar(uint8_t ch)
-{
-	return 0;
-}
-
-void qemu_lf()
-{
-}
-
-void qemu_putinthex(const uint32_t n)
-{
-}
-
-void qemu_puts(const char *s)
-{
-}
-
-void qemu_puthex(uint8_t ch)
-{
-}
-
-void qemu_hexdump(uint8_t *buf, int len)
-{
-}
-#else
-static volatile uint8_t *debugtx = (volatile uint8_t *)TK1_MMIO_QEMU_DEBUG;
-
-int qemu_putchar(uint8_t ch)
-{
-	*debugtx = ch;
-
-	return ch;
-}
-
-void qemu_lf()
-{
-	qemu_putchar('\n');
-}
-
-char hexnibble(uint8_t ch)
-{
-	switch (ch) {
-	case 0x0:
-		return '0';
-	case 0x1:
-		return '1';
-	case 0x2:
-		return '2';
-	case 0x3:
-		return '3';
-	case 0x4:
-		return '4';
-	case 0x5:
-		return '5';
-	case 0x6:
-		return '6';
-	case 0x7:
-		return '7';
-	case 0x8:
-		return '8';
-	case 0x9:
-		return '9';
-	case 0xa:
-		return 'a';
-	case 0xb:
-		return 'b';
-	case 0xc:
-		return 'c';
-	case 0xd:
-		return 'd';
-	case 0xe:
-		return 'e';
-	case 0xf:
-		return 'f';
-	}
-
-	return '0';
-}
-
-void qemu_puthex(uint8_t ch)
-{
-	qemu_putchar(hexnibble(ch >> 4 & 0x0f));
-	qemu_putchar(hexnibble(ch & 0x0f));
-}
-
-void qemu_putinthex(const uint32_t n)
-{
-	uint8_t buf[4];
-
-	memcpy(buf, &n, 4);
-	qemu_puts("0x");
-	for (int i = 3; i > -1; i--) {
-		qemu_puthex(buf[i]);
-	}
-}
-
-void qemu_puts(const char *s)
-{
-	while (*s)
-		qemu_putchar(*s++);
-}
-
-void qemu_hexdump(uint8_t *buf, int len)
-{
-	uint8_t *row;
-	uint8_t *byte;
-	uint8_t *max;
-
-	row = buf;
-	max = &buf[len];
-
-	for (byte = 0; byte != max; row = byte) {
-		// Offset
-		// printf("%07x ", row - buf);
-
-		for (byte = row; byte != max && byte != (row + 16); byte++) {
-			qemu_puthex(*byte);
-		}
-
-		qemu_lf();
-	}
-}
-#endif
-
 void *memset(void *dest, int c, unsigned n)
 {
 	uint8_t *s = dest;
