+++ libcommon/proto.c
@@ -1,8 +1,8 @@
 // Copyright (C) 2022, 2023 - Tillitis AB
 // SPDX-License-Identifier: GPL-2.0-only
 
-#include <lib.h>
 #include <proto.h>
+#include <qemu_debug.h>
 #include <tk1_mem.h>
 
 // clang-format off
